% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountDistinctAlong.R
\name{CountDistinctAlong}
\alias{CountDistinctAlong}
\title{Compute number of distinct values encountered along an ordered vector}
\source{
https://stackoverflow.com/questions/25553099/r-dplyr-cumulative-version-of-n-distinct
}
\usage{
CountDistinctAlong(var)
}
\arguments{
\item{var}{Vector of values.}
}
\description{
Compute number of distinct values encountered along an ordered vector
}
\examples{
df <- data.frame(time = 1:20,
  grp = sort(rep(1:5,4)),
  var1 = rep(c('A','B'),10)
)
CountDistinctAlong(df$grp)
}
