% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateStandardDirectoryStructure.R
\name{CreateStandardDirectoryStructure}
\alias{CreateStandardDirectoryStructure}
\title{Function to create the necessary directories and files for an organized analysis}
\usage{
CreateStandardDirectoryStructure(path = ".", dirs = c("assets", "functions",
  "figures", "data", "data_raw", "reports", "rmds", "r_scripts"))
}
\arguments{
\item{path}{Path to parent folder for the analysis}

\item{dirs}{Directories to create}
}
\description{
Function to create the necessary directories and files for an organized analysis
}
\examples{
\dontrun{
CreateStandardDirectoryStructure()
}
}
