% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CumsumUntilThreshold.R
\name{CumsumUntilThreshold}
\alias{CumsumUntilThreshold}
\title{Group rows based on cumulative sum with reset after reaching threshold}
\source{
https://gist.github.com/jgilfillan/23336d0f5bcfffe6a71d0bdd634d023e
}
\usage{
CumsumUntilThreshold(x, threshold)
}
\arguments{
\item{x}{A vector of numbers.}

\item{threshold}{Threshold for resetting cumsum.}
}
\value{
Vector of groupings such that threshold is never exceeded.
}
\description{
Group rows based on cumulative sum with reset after reaching threshold
}
\examples{
testvector <- c(.5, .1, .2, .9, .9, .2, .5)
CumsumUntilThreshold(testvector, 1)
}
