% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindBinMidpoints.R
\name{FindBinMidpoints}
\alias{FindBinMidpoints}
\title{Calculate the midpoints of bins created using cut function}
\source{
https://mcfromnz.wordpress.com/2014/05/23/finding-the-midpoint-when-creating-intervals/
}
\usage{
FindBinMidpoints(x, dp = 2)
}
\arguments{
\item{x}{Vector of bins returned from cut().}

\item{dp}{Number of decimal points to round.}
}
\value{
numeric vector of bin midpoints
}
\description{
Calculate the midpoints of bins created using cut function
}
\examples{
\dontrun{
x <- c(1, 2, 5, 10, 15, 21, 30, 50, 51, 52, 60)
FindBinMidpoints(cut(x, quantile(x), include.lowest = T))
}
}
