% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round2.R
\name{round2}
\alias{round2}
\title{Round half away from zero. R uses "round to nearest even" by default.}
\source{
https://stackoverflow.com/questions/12688717/round-up-from-5-in-r
}
\usage{
round2(x, n = 0)
}
\arguments{
\item{x}{A number.}

\item{n}{Number of decimal places to be used.}
}
\value{
\code{x} rounded to \code{n} digits.
}
\description{
Round half away from zero. R uses "round to nearest even" by default.
}
\examples{
round2(24.5)
round2(-24.5, 0)
}
