% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsePackages.R
\name{UsePackages}
\alias{UsePackages}
\title{Load the namespace of a package (library), or install if necessary}
\source{
My colleague Wil Doane
}
\usage{
UsePackages(..., repos_ = "https://cran.revolutionanalytics.com")
}
\arguments{
\item{...}{Bare (unquoted) package names}
}
\description{
Load the namespace of a package (library), or install if necessary
}
\examples{
\dontrun{
UsePackages(ggplot2, tidyverse)
}
}
